fix xss issue and remove old upload code
