+++ app/controllers/auth/coauth_controller.rb
@@ -71,25 +71,6 @@ module Auth
       false
     end
 
-    SECRET = Rails.application.secrets.secret_key_base
-
-    def configure_asset_access
-      session_valid = 20.years.from_now
-
-      data = SecureRandom.hex(8)
-      digest = OpenSSL::Digest.new('sha256')
-      hmac = OpenSSL::HMAC.hexdigest(digest, SECRET, data)
-
-      cookies[:verified] = {
-        value: "#{data}.#{hmac}",
-        expires: session_valid,
-        secure: USE_SSL,
-        httponly: true,
-        same_site: :none,
-        path: "/"
-      }
-    end
-
     def store_social(uid, provider)
       value = {
         value: {
